+++ ChangeLog.md
@@ -1,3 +1,9 @@
+# 0.2 (2022-03-18)
+
+* Since v0.4.0.0, sop-core contains `Semigroup` and `Monoid` instances for `NP`,
+  and generics-sop contains `productTypeFrom/To`. Using these, this library's
+  content is almost trivial. :)
+
 # 0.1 (2018-06-22)
 
 * First version. Designed for `base >= 4.11`.